Moved action names to new file
